ATE-2 Week 2 (Stacks, Queues, Sorts)
Finished remaining parts of the exercise but there are performance
issues with Randomized Queue.

There is a problem with my current approach in that the backing queue array
gets fragmented with null values as items are dequeued. This leads to
performance constraint busting and complications in code when shrinking array
or fetching random item (which takes a worst-case linear instead of
constant time).

The trick seems to be in exploiting the fact the array order is random so we
can immediately plug the hole after fetching random element by taking the
*last* element from the array and putting it in the hole. I'll investigate this
further and see if this will be sufficient.